components/Network: move includes to src folder
