+++ components/HTTP/src/RESTServer.cpp
@@ -37,7 +37,7 @@
 #include <freertos/task.h>
 #include <iostream>
 
-#include <Network/Network.h>
+#include "Network.h"
 
 #define REST_CHECK(expr, msg)   \
     do {                        \
