+++ components/Network/CMakeLists.txt
@@ -1,4 +1,4 @@
 file(GLOB_RECURSE SOURCES src/*.cpp)
 idf_component_register(SRCS ${SOURCES}
-                       INCLUDE_DIRS include
+                       INCLUDE_DIRS src
                        REQUIRES Util VAINA)
