+++ components/Network/Kconfig
@@ -17,14 +17,14 @@ config RADIO_SLIP_UART
 
 config RADIO_SLIP_TX
     int "GPIO to use for the TX pin"
-    default 8
+    default 21
     range 0 31
     help
         This is the GPIO that is going to be used for the SLIP TX pin.
 
 config RADIO_SLIP_RX
     int "GPIO to use for the RX pin"
-    default 7
+    default 22
     range 0 31
     help
         This is the GPIO that is going to be used for the SLIP RX pin.
