+++ components/VAINA/src/Vaina.h
@@ -26,7 +26,7 @@
 #include <esp_err.h>
 #include <iostream>
 
-#include <Network/WiFi.h>
+#include "WiFi.h"
 
 /**
  *
